% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDesiredFormat.R
\name{GetDesiredFormat}
\alias{GetDesiredFormat}
\title{Combine six metrics into one summary}
\usage{
GetDesiredFormat(trtdat = sexy1_trtkey, plotdat = sexy1_plotkey)
}
\arguments{
\item{trtdat}{One of the field's xxxx_trtkey (e.g. sexy1_trtkey, eusun1_trtkey).}

\item{plotdat}{One of the field's xxxx_plotkey (e.g. sexy1_plotkey, eusun1_plotkey)}
}
\value{
A data frame
}
\description{
Combine six metrics into one summary
}
