% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prye-package.R
\docType{package}
\name{prye-package}
\alias{prye}
\alias{prye-package}
\title{prye: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vanichols/prye}
  \item Report bugs at \url{https://github.com/vanichols/prye/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
